shut down http server gracefully
