+++ package.json
@@ -84,6 +84,7 @@
   "devDependencies": {
     "@earthstar-project/mini-rpc": "^2.1.0",
     "@types/browser-or-node": "^1.3.0",
+    "@types/express": "^4.17.13",
     "@types/node": "^14.14.43",
     "@types/tap": "^14.10.2",
     "@types/tape": "^4.13.0",
@@ -102,6 +103,7 @@
     "chalk": "^4.1.1",
     "chloride": "^2.4.1",
     "concurrency-friends": "^5.0.1",
+    "express": "^4.17.1",
     "fast-deep-equal": "^3.1.3",
     "fast-json-stable-stringify": "^2.1.0",
     "rfc4648": "^1.4.0",
